backend/requirements.txt
alembic==1.16.4
aniso8601==10.0.1
blinker==1.9.0
click==8.2.1
dotenv==0.9.9
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
